refactor: Make the ContinuousFlyoutMetrics class and file name consistent.
